Keyword searches not restricted to the start of the sentence.
Keyword searches previously only filtered the start of the sentence.
Now spraffbot finds any n-gram containing the keyword and can generate a
sentence backward and forward from that n-gram.  Not sure what the
performance will be like on a large corpus.  Also added some
canonicaliser gubbins, as many of the vecs in the system are believed to
be dups.